refactor: simplify reference select code and disable option creation
option creation does not make sense now that we can toggle between static and
reference mode
