+++ src/components/ComponentEditor/inputs/ReferenceMultiSelect.tsx
@@ -1,11 +1,10 @@
 import { SelectableValue } from "@grafana/data";
-import { MultiSelect, InputControl } from "@grafana/ui";
-import { Control, get } from "react-hook-form";
+import { InputControl, MultiSelect } from "@grafana/ui";
+import { get } from "react-hook-form";
 import { useState } from "react";
-import { toOptions } from "./ReferenceSelect";
 
 import { useComponentContext } from "../../../state";
-import { ExportType } from "../../../lib/components";
+import { ReferenceSelectProps, toOptions } from "./ReferenceSelectBase";
 
 const ReferenceMultiSelect = ({
   control,
@@ -13,29 +12,17 @@ const ReferenceMultiSelect = ({
   exportName,
   rules,
   width,
-}: {
-  control: Control<Record<string, any>>;
-  name: string;
-  exportName: ExportType;
-  rules?: Object;
-  width?: number;
-}) => {
+  defaultValue,
+}: ReferenceSelectProps<{ "-reference": string }[]>) => {
   const { components } = useComponentContext();
-  const defaultValue = get(control.defaultValuesRef.current, name);
-  const [value, setValue] = useState<Array<SelectableValue<object>>>(() => {
+  defaultValue = defaultValue ?? get(control.defaultValuesRef.current, name);
+  const [value, setValue] = useState<SelectableValue<object>[]>(() => {
     if (!defaultValue) return [];
-    return defaultValue.map((x: SelectableValue<object>) => {
-      if ("-reference" in x)
-        return { label: x["-reference"], value: x, key: x["-reference"] };
-      else return { label: JSON.stringify(x), value: x };
-    });
+    return defaultValue.map((v) => ({ label: v["-reference"], value: v }));
   });
-  const [customOptions, setCustomOptions] = useState<
-    Array<SelectableValue<any>>
-  >([]);
   const targetComponents = toOptions(
     components.map((x) => x.block),
-    exportName
+    exportName,
   );
   return (
     <InputControl
@@ -43,21 +30,13 @@ const ReferenceMultiSelect = ({
       render={({ field: { onChange, ref, ...field } }) => (
         <MultiSelect
           {...field}
-          options={[...targetComponents, ...customOptions]}
+          options={targetComponents}
           allowCustomValue={true}
           onChange={(v) => {
-            onChange(v.map((x) => x.value));
+            onChange(v.map((e: SelectableValue) => e.value));
             setValue(v);
           }}
           value={value}
-          onCreateOption={(newOption) => {
-            const customValue: SelectableValue<object> = {
-              value: JSON.parse(newOption),
-              label: newOption,
-            };
-            setCustomOptions([...customOptions, customValue]);
-            setValue([...value, customValue]);
-          }}
           width={width}
         />
       )}
